Renamed files, added info to stdout (verbose).
